Don't iterate if the job iterator isn't present in outputs
If orig_execute encountered an error and did not run, job iterator
will be preent in the prompt but not in the outputs. This would
cause a second exception if some other node had an error, making
it look like the problem is in job iterator when it is not.

Fixes #3