Remove executor hack, convert job to list instead
This makes the node pack work on the latest ComfyUI. Swap the JobIterator
node for a JobToList node. Nothing else needs to change.

This does mean that the workflow will be evaluated breadth-first instead
of depth-first, which means intermediate results will be cached in memory
and you won't get any output until the whole job is finished.